refactor: mvoe `ValidationErrorsBag` to common/errors module
