+++ src/common/errors/validation-errors-bag.spec.ts
@@ -1,5 +1,5 @@
 import { InvalidArgumentError } from "@/common/errors/invalid-argument.error";
-import { ValidationErrorsBag } from "@/infra/http/validation/validation-errors-bag.error";
+import { ValidationErrorsBag } from "@/common/errors/validation-errors-bag.error";
 import { option } from "fp-ts";
 
 describe("ValidationErrorsBag", () => {
